Top 98 Healthiest Food
This is an excel project made by me in 14-15 of October 2024. This project is part of my plan on a roadmap to become a Data Analyst.

I've done this analysis myself with help of Copilot to understand certain things such as how to pull some trick on Power Query and also to help me made my finding explanation look pretty.

Thank you.